phase1/requirements.txt
# Core
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.23
alembic>=1.13.0
aiosqlite>=0.19.0
asyncpg>=0.29.0

# Data connectors
yfinance>=0.2.33
finnhub-python>=2.4.19
alpaca-py>=0.13.0
httpx>=0.25.0
aiohttp>=3.9.0

# Data processing
pandas>=2.1.0
numpy>=1.26.0
pytz>=2023.3

# Utilities
python-dotenv>=1.0.0
cachetools>=5.3.0
structlog>=23.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
httpx>=0.25.0
aiosqlite>=0.19.0

# Development
black>=23.11.0
ruff>=0.1.6
mypy>=1.7.0
python-multipart
